+++ controllers/auth/getCurrent.js
@@ -1,10 +1,16 @@
-const getCurrent = async (req, res) => {
-  const { email, subscription } = req.user;
+const { User } = require("../../models/user");
 
-  res.json({
-    email,
-    subscription,
-  });
-};
+const getCurrent = async (req, res) => {
+    const { email } = req.user;
+    console.log(email)
+    const user = await User.findOne({email});
+
+    res.json({
+        name: user.name,
+        email,
+        avatarURL: user.avatarURL,
+        birthday:user.birthday
+    })
+}
 
 module.exports = getCurrent;
