+++ controllers/auth/logout.js
@@ -1,12 +1,12 @@
 const { User } = require("../../models/user");
 
-const logout = async (req, res) => {
-  const { _id } = req.user;
-  await User.findByIdAndUpdate(_id, { token: "" });
+const logout = async(req, res) => {
+    const {_id} = req.user;
+    await User.findByIdAndUpdate(_id, {token: ""});
 
-  res.status(204).json({
-    message: "Logout success",
-  });
-};
+    res.status(204).json({
+        message: "Logout success"
+    })
+}
 
 module.exports = logout;
