+++ controllers/auth/signIn.js
@@ -1,0 +1,41 @@
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
+
+const { User } = require("../../models/user");
+
+const { HttpError } = require("../../helpers");
+
+const { SECRET_KEY } = process.env;
+
+const signIn = async (req, res) => { 
+    const {email, password} = req.body;
+    const user = await User.findOne({email});
+
+    const passwordCompare = await bcrypt.compare(password, user.password);
+    if(!user || !passwordCompare) {
+        throw HttpError(401, "Email or password is wrong");
+    }
+
+    if (!user.verify) {
+        throw HttpError(401, "Email not verified");  
+    }
+
+    const payload = {
+        id: user._id,
+    }
+
+    const token = jwt.sign(payload, SECRET_KEY, { expiresIn: "23h" });
+    await User.findByIdAndUpdate(user._id, {token});
+
+    res.json({
+        token,
+        "user": {
+            "name": user.name,
+            "email": user.email,
+            "avatarURL": user.avatarURL,
+            "birthdate": user.birthdate,
+        }
+    })
+}
+
+module.exports = signIn;
