+++ controllers/auth/verifyEmail.js
@@ -1,0 +1,20 @@
+const { User } = require("../../models/user");
+
+const { HttpError } = require("../../helpers");
+
+const verifyEmail = async (req, res) => {
+    const { verificationToken } = req.params;
+    console.log(verificationToken);
+    
+    const user = await User.findOne({verificationToken});
+    if(!user){
+        throw HttpError(404, "User not found")
+    }
+    await User.findByIdAndUpdate(user._id, {verify: true, verificationToken: ""});
+
+    res.json({
+        message: "Verification successful"
+    })
+}
+
+module.exports = verifyEmail;
